djangorest/requirements.txt
Django==1.11.5
pkg-resources==0.0.0
pytz==2017.2
djangorestframework
Django==1.11.5
pkg-resources==0.0.0
pytz==2017.2
